requirements.txt
alembic==1.16.4
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
crashtest==0.4.1
cryptography==45.0.5
cycler==0.12.1
distlib==0.3.9
dnspython==2.7.0
docopt==0.6.2
dotmap==1.3.30
dulwich==0.22.8
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fonttools==4.59.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
keyring==25.6.0
kiwisolver==1.4.9
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
mjml==0.11.1
more-itertools==10.7.0
MouseInfo==0.1.3
msgpack==1.1.1
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pbs-installer==2025.7.12
pefile==2023.2.7
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
plyer==2.1.0
poetry==2.1.3
poetry-core==2.1.3
psutil==7.0.0
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyGetWindow==0.0.9
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
PyMsgBox==1.0.9
pynput==1.8.1
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
sendgrid==6.12.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.1
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
Werkzeug==3.1.3
win10toast==0.9
win32_setctime==1.2.0
zstandard==0.23.0
